% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{data.simulate}
\alias{data.simulate}
\title{Simulate normally distributed or scRNAseq data based on partial correlation matrix}
\usage{
data.simulate(
  features,
  samples,
  graph = "random",
  degree = 2,
  type = "normal",
  ZI = FALSE,
  depth = 50000
)
}
\arguments{
\item{features}{integer. Number of genes or features}

\item{samples}{integer. Number of cells/samples/observations}

\item{graph}{type of graph or network. Available options: random, hub, cluster, band, scale-free}

\item{degree}{degree of simulated graph or average number of edges per node. Default: 2}

\item{type}{normal or ESCO for normally distributed or scRNAseq data simulation, respectively}

\item{ZI}{TRUE or FALSE (default). Whether simulating zero-inflated count or not in scRNAseq data simulation}

\item{depth}{integer, sequencing depth. For scRNAseq, 20 000 reads are recommended by most workflows. Default: 50 000 reads}
}
\value{
list of simulated partial correlation matrix (pcor.sim) and data matrix (sim.dat)
}
\description{
Using partial correlation matrix to simulate data
}
