% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{escoSimulateSingleGroup}
\alias{escoSimulateSingleGroup}
\title{Simulate homogeneous scRNAseq data from ESCO model}
\usage{
escoSimulateSingleGroup(
  GeneNumbers,
  CellNumbers,
  pcorSim,
  depth = 50000,
  zeroprop = 0.4,
  zisim = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{GeneNumbers}{integer, number of genes to be simulated}

\item{CellNumbers}{integer, number of homogeneous cells to be simulated}

\item{pcorSim}{partial correlation matrix which contains information of networks. Non-zero value equals to presence of edge between two genes.}

\item{depth}{integer, sequencing depth. For scRNAseq, 20 000 reads are recommended by most workflows. Default: 50 000 reads}

\item{zeroprop}{value from 0 to 1, maximum proportion of zero counts of genes in network (pcorSim). Default: 0.4}

\item{zisim}{TRUE or FALSE. Simulating zero-inflated counts from model of ESCO R package}

\item{verbose}{TRUE or FALSE. Information of simulating process}
}
\value{
list of simulated count (genes x cells) and partial correlation matrix.
}
\description{
Simulate a simple model of single-cell RNA sequencing (scRNAseq) data
with cells from the same population and no outlier cells (homogeneous) based on ESCO model (ESCO R package)
}
