% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_sim.R
\name{graph.generator}
\alias{graph.generator}
\title{Network simulation}
\usage{
graph.generator(
  d = NULL,
  graph = "random",
  degree = 2,
  g = NULL,
  prob = NULL,
  pcor.return = TRUE
)
}
\arguments{
\item{d}{number of nodes in network}

\item{graph}{type of network structure. Five options: random (default), hub, cluster, band, scale-free}

\item{degree}{average number of edges per node in random graph. Default: 2}

\item{g}{number of hubs or clusters in hub and cluster graphs, respectively.}

\item{prob}{parameter for cluster graph, probability that a pair of nodes has an edge in each cluster. Default: 6*g/d}

\item{pcor.return}{TRUE or FALSE. return only partial correlation matrix or a list of all relevant simulated matrices.}
}
\value{
return partial correlation matrix by default
}
\description{
Simulating different network structure (random, hub, cluster, band, scale-free). Simulation of hub, cluster and band follow models from huge.generator (huge R package)
}
