% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_sim.R
\name{simulateBAmodel}
\alias{simulateBAmodel}
\title{Simulate scale-free network following Barabasi-Albert model}
\usage{
simulateBAmodel(num_of_nodes, m)
}
\arguments{
\item{num_of_nodes}{integers. Number of nodes or features in network}

\item{m}{can be floating numbers, lambda value of Poisson distribution which simulates k, number of iterations for preferential attachments}
}
\value{
adjacency matrix
}
\description{
Simulate scale-free network using Barabasi-Albert model from https://doi.org/10.1103/RevModPhys.74.47
with two important properties. Growth and preferential attachment
}
